fix(import): refactoring currency
