+++ apps/next/src/pages/Settings/Settings.tsx
@@ -1,17 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import {
-  Typography,
-  Avatar,
-  Button,
-  MenuItem,
-  Select,
-  Stack,
-  Switch,
-} from '@avalabs/k2-alpine';
-
+import { Button, Stack, Switch } from '@avalabs/k2-alpine';
 import { useSettingsContext } from '@core/ui';
-
 import {
   BUG_BOUNTIES_URL,
   CORE_FEATURE_REQUEST_URL,
@@ -29,12 +19,11 @@ import {
   SettingsCard,
   Footer,
 } from './components';
-import { currencies } from '@core/types';
-import { runtime } from 'webextension-polyfill';
+import { CurrencySelector } from './components/CurrencySelector';
 
 export const Settings = () => {
   const { t } = useTranslation();
-  const { lockWallet, updateCurrencySetting, currency } = useSettingsContext();
+  const { lockWallet } = useSettingsContext();
 
   const [isTestnetMode, setIsTestnetMode] = useState(false);
   const [isPrivacyMode, setIsPrivacyMode] = useState(false);
@@ -82,64 +71,7 @@ export const Settings = () => {
         <SettingsNavItem
           divider
           label={t('Currency')}
-          secondaryAction={
-            // this needs to be replaced with SelectCountry when it is ready in k2-alpine
-            <Select
-              label={t('Currency')}
-              value={currency}
-              renderValue={(selected) => {
-                const selectedCurrency = currencies.find(
-                  (c) => c.symbol === selected,
-                );
-                if (!selectedCurrency) {
-                  return;
-                }
-                const countryCode = selectedCurrency.countryCode;
-
-                return (
-                  <Stack
-                    direction="row"
-                    alignItems="center"
-                    justifyContent="flex-end"
-                    gap={1}
-                  >
-                    <Avatar
-                      sx={{
-                        width: '16px',
-                        height: '16px',
-                      }}
-                      src={runtime.getURL(
-                        `images/currencies/${countryCode.toLowerCase()}.svg`,
-                      )}
-                      alt={`${countryCode} flag`}
-                      slotProps={{
-                        img: {
-                          loading: 'lazy',
-                          sx: {
-                            objectFit: 'cover',
-                          },
-                        },
-                      }}
-                    />
-                    <Typography>{selectedCurrency.symbol}</Typography>
-                  </Stack>
-                );
-              }}
-              onChange={(e) => {
-                const newValue = e.target.value;
-                const found = currencies.find((c) => c.symbol === newValue);
-                if (found) {
-                  updateCurrencySetting(found.symbol);
-                }
-              }}
-            >
-              {currencies.map((c) => (
-                <MenuItem key={c.symbol} value={c.symbol}>
-                  {`${c.label} (${c.symbol})`}
-                </MenuItem>
-              ))}
-            </Select>
-          }
+          secondaryAction={<CurrencySelector />}
         />
         <SettingsNavItem
           divider
